Move `atl` into param of `/janet`
