+++ app.py
@@ -1,11 +1,10 @@
-
 import os
 from datetime import date
 
 import dotenv
 from flask import Flask, request
 
-from janet.functions import say, valid
+from janet.functions import say, valid, russ_travel
 
 dotenv_path = os.path.join(os.path.dirname(__file__), '.env')
 dotenv.load_dotenv(dotenv_path)
@@ -16,11 +15,16 @@ app = Flask(__name__)
 @app.route("/bot/janet", methods=['POST'])
 def janet():
     if valid(request):
-        return say(
-            "Hi, there!",
-            "http://78.media.tumblr.com/daf862b49b82e49a47354b14c5143363/"
-            "tumblr_oefwev7pLA1vvi3bvo8_250.gif"
-        )
+        text = request.form['text']
+        if text.lower() == 'atl':
+            response = russ_travel()
+        else:
+            response = say(
+                "Hi, there!",
+                "http://78.media.tumblr.com/daf862b49b82e49a47354b14c5143363/"
+                "tumblr_oefwev7pLA1vvi3bvo8_250.gif"
+            )
+        return response
 
 
 @app.route("/bot/badjanet", methods=['POST'])
@@ -33,23 +37,23 @@ def hello_dinks():
         )
 
 
-@app.route("/bot/atl", methods=['POST'])
-def russ_travel():
-    if valid(request):
-        d0 = date.today()
-        d1 = date(2019, 6, 13)
-        delta = d1 - d0
-        if delta.days > 0:
-            return say(
-                f'{delta.days} days until Russ moves back!',
-                'https://media.giphy.com/media/xULW8poAH4m1KBQr1C/giphy.gif'
-            )
-        else:
-            return say(
-                'Russ is already back!',
-                'https://media.giphy.com/media/xUOxeRRkTYdQJfyy2Y/giphy.gif'
-            )
-
-
-if __name__ == "__main__":
-    app.run()
+# @app.route("/bot/atl", methods=['POST'])
+# def russ_travel():
+#     if valid(request):
+#         d0 = date.today()
+#         d1 = date(2019, 6, 13)
+#         delta = d1 - d0
+#         if delta.days > 0:
+#             return say(
+#                 f'{delta.days} days until Russ moves back!',
+#                 'https://media.giphy.com/media/xULW8poAH4m1KBQr1C/giphy.gif'
+#             )
+#         else:
+#             return say(
+#                 'Russ is already back!',
+#                 'https://media.giphy.com/media/xUOxeRRkTYdQJfyy2Y/giphy.gif'
+#             )
+#
+#
+# if __name__ == "__main__":
+#     app.run()
