+++ janet/__init__.py
@@ -1,0 +1,1 @@
+from .functions import say, valid, russ_travel
