+++ janet/functions.py
@@ -1,5 +1,6 @@
 
 import os
+from datetime import date
 from flask import jsonify
 
 VERIFICATION_TOKEN = os.getenv('VERIFICATION_TOKEN', 'sesame')
@@ -16,3 +17,19 @@ def say(response, image_url=None):
 def valid(req):
     # TODO: Switch to secrets
     return req.form['token'] == VERIFICATION_TOKEN
+
+
+def russ_travel():
+    d0 = date.today()
+    d1 = date(2019, 6, 13)
+    delta = d1 - d0
+    if delta.days > 0:
+        return say(
+            f'{delta.days} days until Russ moves back!',
+            'https://media.giphy.com/media/xULW8poAH4m1KBQr1C/giphy.gif'
+        )
+    else:
+        return say(
+            'Russ is already back!',
+            'https://media.giphy.com/media/xUOxeRRkTYdQJfyy2Y/giphy.gif'
+        )
